+++ NEWS.md
@@ -1,3 +1,5 @@
+# rules (development version)
+
 # rules 0.1.0
 
 * The default encoding methods were changed such that `parsnip` will respect the encoding that the underlying model uses.